requirements.txt
discord.py==2.0.1
python-dotenv==0.21.0
requests==2.26.0
qrcode==7.3
```